% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageseird_Methods.R
\name{calculateCurrentState,ageseirdModel-method}
\alias{calculateCurrentState,ageseirdModel-method}
\title{An S4 method to calculate current state of an object of the ageseirdModel class}
\usage{
\S4method{calculateCurrentState}{ageseirdModel}(epiModel, deaths, nderiv = 8, plotDeriv = FALSE)
}
\arguments{
\item{epiModel}{An age-SEIRD model, whose current state is to be calculated.}

\item{deaths}{The total death count for this model, up to each of the
changeTimes so far.}

\item{nderiv}{Which derivatives of D to assume is sum to 0.}

\item{plotDeriv}{True/False that tells the function to plot the derivative
of D determined by nderiv.}
}
\value{
An ageseirdModel of object.
}
\description{
Calls two generic functions to calculate the downstream nodes and estimate
the infectious population in turn.
}
\examples{
#set up model
model <- setAgeSIRD(N = c(100,100), Betas = matrix(c(1,0.5,0.2,1), nrow=2, byrow=TRUE),
Gamma = 1/5, Lambda = 1/2, ProbOfDeath = 1/20, I0 = c(1,1))
#Set the deaths
deaths <- matrix(c(10,20), nrow = 1, ncol = 2)
time <- 10
#calculate current state
model <- calculateCurrentState(model, time, deaths)
#return the current state
currentState(model)
}
