% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agesird_Methods.R
\name{calculateCurrentState,agesirdModel-method}
\alias{calculateCurrentState,agesirdModel-method}
\title{An S4 method to calculate current state of an object of the agesirdModel class}
\usage{
\S4method{calculateCurrentState}{agesirdModel}(epiModel, deaths)
}
\arguments{
\item{epiModel}{An agesird model, whose current state is to be calculated.}

\item{deaths}{The total death count for each age-group in this model,
up to each of the changeTimes so far. Rows relate to age groups and columns
to changeTimes.}
}
\value{
An agesirdModel object.
}
\description{
Calls two generic functions to calculate the downstream nodes and estimate
the infectious population in turn.
}
\examples{
#set up model
model <- setAgeSIRD(N = c(100,100), Betas = matrix(c(1,0.5,0.2,1), nrow=2, byrow=TRUE),
                    Gamma = 1/5, ProbOfDeath = 1/20, I0 = c(1,1))
#Set the deaths
deaths <- c(10,20)
time <- 10
#calculate current state
model <- calculateCurrentState(model, time, deaths)
#return the current state
currentState(model)
}
