% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiModel_Methods.R
\name{calculateCurrentState,epiModel-method}
\alias{calculateCurrentState,epiModel-method}
\title{An S4 method to calculate current state of an object of the epiModel class}
\usage{
\S4method{calculateCurrentState}{epiModel}(epiModel, ...)
}
\arguments{
\item{epiModel}{The epidemic model, whose current state is to be calculated.}
}
\value{
An epiModel of object of the same class as that given.
}
\description{
Calls two generic functions to calculate the downstream nodes and esimate
the infectious population in turn. These will be specific to each type of
epiModel class.
}
\examples{
#to do
}
