% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seird_Methods.R
\name{calculateCurrentState,seirdModel-method}
\alias{calculateCurrentState,seirdModel-method}
\title{An S4 method to calculate current state of an object of the seirdModel class}
\usage{
\S4method{calculateCurrentState}{seirdModel}(epiModel, deaths, nderiv = 6, plotDeriv = FALSE)
}
\arguments{
\item{epiModel}{An SEIRD model, whose current state is to be calculated.}

\item{deaths}{The total death count for this model, up to each of the
changeTimes so far.}

\item{nderiv}{Which derivative of D to assume is 0.}

\item{plotDeriv}{True/False that tells the function to plot the dervivative
of D determined by nderiv.}
}
\value{
An seirdModel of object.
}
\description{
Calls two generic functions to calculate the downstream nodes and estimate
the infectious population in turn.
}
\examples{
#set up model
model <- setSEIRD(N = 100, Beta = 1, Gamma = 1/5, Lambda = 1/3, ProbOfDeath = 0.5, I0 = 1)
#Set the deaths
deaths <- 20
time <- 10
#calculate current state
model <- calculateCurrentState(model, time, deaths)
#return the current state
currentState(model)
}
