% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sird_Methods.R
\name{calculateCurrentState,sirdModel-method}
\alias{calculateCurrentState,sirdModel-method}
\title{An S4 method to calculate current state of an object of the sirdModel class}
\usage{
\S4method{calculateCurrentState}{sirdModel}(epiModel, deaths)
}
\arguments{
\item{epiModel}{An sird model, whose current state is to be calculated.}

\item{deaths}{The total death count for this model, up to each of the
changeTimes so far.}
}
\value{
An sirdModel object.
}
\description{
Calls two generic functions to calculate the downstream nodes and estimate
the infectious population in turn.
}
\examples{
#set up model
model <- setSIRD(N = 100, Beta = 1, Gamma = 1/5, ProbOfDeath = 0.5, I0 = 1)
#Set the deaths
deaths <- 20
time <- 10
#calculate current state
model <- calculateCurrentState(model, time, deaths)
#return the current state
currentState(model)
}
