% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agesird_Methods.R
\name{calculateDownstreamExponentialNodes,agesirdModel-method}
\alias{calculateDownstreamExponentialNodes,agesirdModel-method}
\title{An S4 method to calculate current state of R and D for an ageSIRD/ageSEIRD when
provided with a count of the total number of deaths for each age-group.}
\usage{
\S4method{calculateDownstreamExponentialNodes}{agesirdModel}(epiModel, deaths)
}
\arguments{
\item{epiModel}{The epidemic model of class age-SIRD to have the current
state of R and D calculated.}

\item{deaths}{The total death count for this model, so far.}
}
\value{
An object of class agesirdModel with the values for D and
R updated for the current state.
}
\description{
Please note that this method does not guarantee that this state would be
generated by the model, given its initial conditions.
Calculates the number of people in the recovered state by scaling D by the
rate of recovery divided by the rate of death.
These values are then stored in the list in the slot "currentState".
}
\examples{
#create example model
model <- setAgeSIRD(N = c(100,100), Betas = matrix(c(1,0.5,0.2,1), nrow=2, byrow=TRUE),
                    Gamma = 1/5, ProbOfDeath = 1/20, I0 = c(1,1))
#Set the deaths (only one vector is required)
deaths <- matrix(c(20,50),byrow=TRUE,ncol = 1)
#call this function
model <- excalibur::calculateDownstreamExponentialNodes(model, deaths=deaths)

currentState(model)
}
