% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sird_Methods.R
\name{calculateDownstreamExponentialNodes,sirdModel-method}
\alias{calculateDownstreamExponentialNodes,sirdModel-method}
\title{An S4 method to calculate current state of R and D for an SIRD when provided
with a count of the total number of deaths.}
\usage{
\S4method{calculateDownstreamExponentialNodes}{sirdModel}(epiModel, deaths)
}
\arguments{
\item{epiModel}{The epidemic model of class SIRD to have the current state of
R and D calculated.}

\item{deaths}{The total death count for this model, so far.}
}
\value{
An object of class sirdModel with the values for D and R updated
for the current state.
}
\description{
Please note that this method does not guarantee that this state would be
generated by the model, given its initial conditions.
Calculates the number of people in the recovered state by scaling D by the
rate of recovery divided by the rate of death.
These values are then stored in the list in the slot "currentState".
}
\examples{
#create example model
model <- setSIRD(N = 100, Beta = 1, Gamma = 1/5, ProbOfDeath = 0.5, I0 = 1)
#Set the deaths (only one entry is required)
deaths <- 20
#call this function
model <- excalibur::calculateDownstreamExponentialNodes(model, deaths=deaths)

currentState(model)
}
