% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentState_Generics.R
\name{calculateDownstreamExponentialNodes}
\alias{calculateDownstreamExponentialNodes}
\title{A inner function to calculate the states of downstream exponential model
compartments.}
\usage{
calculateDownstreamExponentialNodes(epiModel, ...)
}
\arguments{
\item{epiModel}{The epidemic model, whose current state is to be calculated.}

\item{...}{Any other arguements to provide to the method.}
}
\description{
This calculates, for example R and D in an SIRD model.
}
