% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiModel_Methods.R
\name{currentState,epiModel-method}
\alias{currentState,epiModel-method}
\title{An S4 method to return current state of an object of the epiModel class}
\usage{
\S4method{currentState}{epiModel}(epiModel)
}
\arguments{
\item{epiModel}{The epidemic model, whose current state is to be returned.}
}
\value{
A list of labelled values.
}
\description{
An S4 method to return current state of an object of the epiModel class
}
\examples{
#set up model
model <- setSIR(N = 100, Beta = 1, Gamma = 1/5, ProbOfDeath = 0.5, I0 = 1)
#Set the deaths
deaths <- 20
time <- 10
#calculate current state
model <- calculateCurrentState(model, time, deaths)
#return the current state
currentState(model)
}
