% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentState_Generics.R
\name{currentState}
\alias{currentState}
\title{A function to return the list of values in the currentState slot of an
epimodel object.}
\usage{
currentState(epiModel, ...)
}
\arguments{
\item{epiModel}{The epidemic model, whose current state is to be print.}

\item{...}{Any other arguements to provide to the method.}
}
\description{
A function to return the list of values in the currentState slot of an
epimodel object.
}
