% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agesird_Methods.R
\name{estimateInfectiousNode,agesirdModel-method}
\alias{estimateInfectiousNode,agesirdModel-method}
\title{An S4 method to estimate current state of S and I for an age-SIRD when provided
with a total count of deaths so far}
\usage{
\S4method{estimateInfectiousNode}{agesirdModel}(epiModel, deaths)
}
\arguments{
\item{epiModel}{The epidemic model of class age-SIRD to have the current state of
S and I estimated.}

\item{deaths}{The total death count for this model, up to each of the
changeTimes so far. Rows relate to changeTimes and columns to age groups.}
}
\value{
An object of class agesirdModel with the values for S and I updated
for the current state.
}
\description{
Please note that this method does not guarantee that this state would be
generated by the model, given its initial conditions.
Due to rounding etc, when the value of I would be very small the estimated I
will be quite different from the one generated when the model is simulated.
These values are then stored in the list in the slot "currentState".
}
\examples{
#create example model
model <- setAgeSIRD(N = c(100,100), Betas = matrix(c(1,0.5,0.2,1), nrow=2, byrow=TRUE),
                    Gamma = 1/5, ProbOfDeath = 1/20, I0 = c(1,1))
#Set the deaths (only one entry is required)
deaths <- matrix(c(20,50),byrow=TRUE, ncol = 1)
#Set the time, makes no difference to this calculation, only used for simulate
time <- 10
#call this + the D+R node function
model <- calculateCurrentState(model, time, deaths)

#check S and I
currentState(model)

#model with time-varying Beta
Betas <- array(NA, dim=c(2,2,2)) #set up array, the first dimension will be time
Betas[1,,] <- matrix(c(1,0.5,
                       0.2,1), nrow=2, byrow = TRUE)
Betas[2,,] <- matrix(c(1,0.1,
                       0,5), nrow=2, byrow = TRUE)
model <- setAgeSIRD(N = c(100,100), Beta = Betas,
                    Gamma = 1/5, ProbOfDeath = 1/20, I0 = c(1,1), changeTimes = 5)
#Set the deaths, two columns are required now
deaths <- matrix(c(10,20,
                   30,50),byrow=TRUE, ncol = 2)
time <- 10
#the model assumes that the first number 20 is the number of deaths up to the
#specified change time 5, and that the last number 30 is the number of deaths
#up to the current time (10).
#if the current state is not past 5, then we would set the final column identical
#to the one before

#call this + the D+R node function
model <- calculateCurrentState(model, t=time, deaths=deaths)

#check S and I
currentState(model)

}
