% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_Methods.R
\name{estimateInfectiousNode,sirModel-method}
\alias{estimateInfectiousNode,sirModel-method}
\title{An S4 method to estimate current state of S and I for an SIR when provided
with a time-series of deaths}
\usage{
\S4method{estimateInfectiousNode}{sirModel}(epiModel, deaths)
}
\arguments{
\item{epiModel}{The epidemic model of class SIR to have the current state of
S and I estimated.}

\item{deaths}{A time-series of the cumulative death counts for this model.}
}
\value{
An object of class sirModel with the values for S and I updated
for the current state.
}
\description{
Please note that this method does not guarantee that this state would be
generated by the model, given its initial conditions.
At the present this model assumes a difference equation set-up to calculate
the number of infected/susceptible and so will be close to but not the same
as the result from an equivalent differential equations model.
These values are then stored in the list in the slot "currentState".
}
\examples{
#create example model
model <- setSIR(N = 100, Beta = 1, Gamma = 1/5, ProbOfDeath = 0.5, I0 = 1)
#Set the deaths (only one entry is required)
deaths <- c(10,20)

#call the D+R node function
model <- excalibur::calculateDownstreamExponentialNodes(model, deaths=deaths)

#call this function

}
