% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excalibur.R
\docType{package}
\name{excalibur}
\alias{excalibur}
\title{excalibur: A package for looking at the calibrations in ODE epidemic models}
\description{
The excalibur package is work in progress
}
\section{excalibur functions}{

setSIR: A function that produces a model of the SIR class in odin
}

