% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seird_Class.R
\name{setSEIRD}
\alias{setSEIRD}
\title{Produce an SEIRD with time-varying Beta values}
\usage{
setSEIRD(N, Beta, Lambda, Gamma, ProbOfDeath, I0, changeTimes = NULL)
}
\arguments{
\item{N}{Total population.}

\item{Beta}{Transmission parameter(s).}

\item{Lambda}{Rate of transition from exposed to infectious.}

\item{Gamma}{Rate of recovery.}

\item{ProbOfDeath}{Instantaneous probability of death whilst infected.}

\item{I0}{Size of initial infected population.}

\item{changeTimes}{If length(Beta)>1 then the corresponding time-points at
which the parameters are changed.}
}
\value{
A object of class sirdModel with odin model code for the same model,
and filled slots for parameters and the initial state.
}
\description{
First section of the function is a wrapper for the internal generator
function for the class sierdModel. The second part setups up the Odin model in
the odinModel slot with the specified parameter values.
}
\examples{
#Standard SIRD with no deaths
model <- setSEIRD(N = 10, Beta = 1, Lambda = 1/2, Gamma = 1/5, ProbOfDeath = 0, I0 = 1)

#SIRD with a 50\% probability of death an any time whilst infected
model <- setSEIRD(N = 10, Beta = 1, Lambda = 1/2, Gamma = 1/5, ProbOfDeath = 0.5, I0 = 1)

#SIRD with time-varying Beta at 10 and 50 days
model <- setSEIRD(N = 10, Beta = c(5,1,3), Lambda = 1/2, Gamma = 1/5, ProbOfDeath = 0.5,
I0 = 1, changeTimes = c(10, 50))
}
