% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirClass.R
\name{setSIR}
\alias{setSIR}
\title{Produce an SIR with time-varying Beta values}
\usage{
setSIR(N, Beta, Gamma, ProbOfDeath, I0, changeTimes = NULL)
}
\arguments{
\item{N}{Total population.}

\item{Beta}{Transmission parameter(s).}

\item{Gamma}{Rate of recovery.}

\item{ProbOfDeath}{Instantaneous probability of death whilst infected.}

\item{I0}{Size of initial infected population.}

\item{changeTimes}{If length(Beta)>1 then the corresponding time-points at
which the parameters are changed.}
}
\description{
First section of the function is a wrapper for the internal generator
function for the class sirModel. The second part setups up the Odin model in
the odinModel slot with the specified parameter values.
}
