% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiModel_Methods.R
\name{calculateCurrentState,epiModel-method}
\alias{calculateCurrentState,epiModel-method}
\title{An S4 method to calculate current state of an object of the epiModel class}
\usage{
\S4method{calculateCurrentState}{epiModel}(epiModel, t, ...)
}
\arguments{
\item{epiModel}{The epidemic model, whose current state is to be calculated.}

\item{t}{The time this state relates to, if the initial conditions occured at
0.}

\item{...}{Any other arguments to provide to the method.}
}
\value{
An epiModel of object of the same class as that given.
}
\description{
Calls two generic functions to calculate the downstream nodes and estimate
the infectious population in turn. These will be specific to each type of
epiModel class.
}
\examples{
#set up model
model <- setSIRD(N = 100, Beta = 1, Gamma = 1/5, ProbOfDeath = 0.5, I0 = 1)
#Set the deaths
deaths <- 20
time <- 10
#calculate current state
model <- calculateCurrentState(model, time, deaths)
#return the current state
currentState(model)
}
