% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentState_Generics.R
\name{calculateCurrentState}
\alias{calculateCurrentState}
\title{A generic function to call the relevant methods to calculate the current
state at a given time based upon some inputs}
\usage{
calculateCurrentState(epiModel, ...)
}
\arguments{
\item{epiModel}{The epidemic model, whose current state is to be calculated.}

\item{...}{Any other arguements to provide to the method.}
}
\description{
At present only has one method, which is for all objects of class "epiModel".
}
