% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seird_Methods.R
\name{estimateInfectiousNode,seirdModel-method}
\alias{estimateInfectiousNode,seirdModel-method}
\title{An S4 method to estimate current state of S, E and I for an SEIRD when provided
with a total count of deaths so far}
\usage{
\S4method{estimateInfectiousNode}{seirdModel}(epiModel, deaths, nderiv)
}
\arguments{
\item{epiModel}{The epidemic model of class SEIRD to have the current state of
S and I estimated.}

\item{deaths}{The total death count for this model, up to each of the
changeTimes so far.}

\item{nderiv}{Which derivative to assume is 0.}
}
\value{
An object of class seirdModel with the values for S, E and I updated
for the current state.
}
\description{
Please note that this method does not guarantee that this state would be
generated by the model, given its initial conditions.
This method uses the methods for an SIRD model to calculate S and R
This method also makes the assumption that the nth derivative of D with
respect to time is 0. N is set by nderiv with a default of 8.
This equation is then solved for I using optim, hence this method returns a
close but non-exact solution for E and I.
These values are then stored in the list in the slot "currentState".
}
\examples{
#model with time-varying Beta
model <- setSEIRD(N = 100, Beta = c(2,1/2), Lambda = 1, Gamma = 1/5, ProbOfDeath = 0.5,
 I0 = 1, changeTimes = 5)
#Set the deaths, two entries are required now
deaths <- c(20,30)
time <- 10

#call this method via calculate current state
model <- calculateCurrentState(model, t=time, deaths=deaths, nderiv = 8)

#check S and I
currentState(model)

}
