% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_Methods.R
\name{estimateInfectiousNode,sirModel-method}
\alias{estimateInfectiousNode,sirModel-method}
\title{An S4 method to estimate current state of S and I for an SIR when provided
with a total count of deaths so far}
\usage{
\S4method{estimateInfectiousNode}{sirModel}(epiModel, deaths)
}
\arguments{
\item{epiModel}{The epidemic model of class SIR to have the current state of
S and I estimated.}

\item{deaths}{The total death count for this model, so far.}
}
\value{
An object of class sirModel with the values for S and I updated
for the current state.
}
\description{
Please note that this method does not guarantee that this state would be
generated by the model, given its initial conditions.
Due to rounding etc, when the value of I would be very small the esimated I
will be quite different from the one generated when the model is simulated.
At the present time-varying Betas are not adjusted for.
These values are then stored in the list in the slot "currentState".
}
\examples{
#create example model
model <- setSIR(N = 100, Beta = 1, Gamma = 1/5, ProbOfDeath = 0.5, I0 = 1)
#Set the deaths (only one entry is required)
deaths <- 20

#call this + the D+R node function
model <- calculateCurrentState(model, deaths=deaths)

#check S and I
currentState(model)
}
