% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentState_Generics.R
\name{estimateInfectiousNode}
\alias{estimateInfectiousNode}
\title{A inner function to calculate the states ofnon-exponential model
compartments.}
\usage{
estimateInfectiousNode(epiModel, ...)
}
\arguments{
\item{epiModel}{The epidemic model, whose current state is to be calculated.}

\item{...}{Any other arguements to provide to the method.}
}
\description{
This calculates, for example S and I in an SIRD model.
}
