% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirClass.R
\name{setSIR}
\alias{setSIR}
\title{Produce an SIR with time-varying Beta values}
\usage{
setSIR(N, Beta, Gamma, ProbOfDeath, I0, changeTimes = NULL)
}
\arguments{
\item{N}{Total population.}

\item{Beta}{Transmission parameter(s).}

\item{Gamma}{Rate of recovery.}

\item{ProbOfDeath}{Instantaneous probability of death whilst infected.}

\item{I0}{Size of initial infected population.}

\item{changeTimes}{If length(Beta)>1 then the corresponding time-points at
which the parameters are changed.}
}
\value{
A object of class sirModel with odin model code for the same model.
}
\description{
First section of the function is a wrapper for the internal generator
function for the class sirModel. The second part setups up the Odin model in
the odinModel slot with the specified parameter values.
}
\examples{
#Standard SIR with no deaths
model <- setSIR(N = 10, Beta = 1, Gamma = 1/5, ProbOfDeath = 0, I0 = 1)

#SIR with a 50\% probability of death an any time whilst infected
model <- setSIR(N = 10, Beta = 1, Gamma = 1/5, ProbOfDeath = 0.5, I0 = 1)

#SIR with time-varying Beta at 10 and 50 days
model <- setSIR(N = 10, Beta = c(5,1,3), Gamma = 1/5, ProbOfDeath = 0.5,
I0 = 1, changeTimes = c(10, 50))
}
