% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiModel_Methods.R
\name{simulate,epiModel-method}
\alias{simulate,epiModel-method}
\title{An S4 method run a realisation of an object of the epiModel class}
\usage{
\S4method{simulate}{epiModel}(object, t, useCurrent = FALSE, tcrit = NULL)
}
\arguments{
\item{object}{An object of the epiModel class.}

\item{t}{The times at which to simulate from the given epiModel object, the
first time is always taken as the start time.}

\item{useCurrent}{True or False, tells the function to simulate from the
current state.}

\item{tcrit}{Critical times (?).}
}
\value{
A dataframe of the odin models run output.
}
\description{
Essentially a wrapper for the run R6 method for the odin model. Forces the
model to start from 0, to keep it consistent. Also puts the output into a
data frame, for easier use.
}
\examples{
#set up our model
model <- setSIRD(N = 10, Beta = 1, Gamma = 1/5, ProbOfDeath = 0, I0 = 1)
#run the model with simulate
simulate(model, t = seq(1,10))
}
